changed medication strengths to being a double instead of an int
