+++ app/femr/data/models/mysql/concepts/ConceptDiagnosis.java
@@ -16,31 +16,30 @@
      along with fEMR.  If not, see <http://www.gnu.org/licenses/>. If
      you have any questions, contact <info@teamfemr.org>.
 */
-package femr.data.models.mysql;
+package femr.data.models.mysql.concepts;
 
-import femr.data.models.core.IConceptMedicationUnit;
+import femr.data.models.core.IConceptDiagnosis;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 @Entity
-@Table(name = "concept_medication_units")
-public class ConceptMedicationUnit implements IConceptMedicationUnit {
+@Table(name = "concept_diagnoses")
+public class ConceptDiagnosis implements IConceptDiagnosis {
+
     @Id
     @Column(name = "id", unique = true, nullable = false)
     private int id;
-    @Column(name = "name", unique = true, nullable = false)
+    @Column(name = "name", nullable = false)
     private String name;
-    @Column(name="description")
-    private String description;
-    @Column(name="isDeleted")
-    private Boolean isDeleted;
 
     @Override
     public int getId() {
         return id;
+    }
+
+    @Override
+    public void setId(int id) {
+        this.id = id;
     }
 
     @Override
@@ -51,25 +50,5 @@ public class ConceptMedicationUnit implements IConceptMedicationUnit {
     @Override
     public void setName(String name) {
         this.name = name;
-    }
-
-    @Override
-    public String getDescription() {
-        return description;
-    }
-
-    @Override
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    @Override
-    public Boolean getIsDeleted() {
-        return isDeleted;
-    }
-
-    @Override
-    public void setIsDeleted(Boolean isDeleted) {
-        this.isDeleted = isDeleted;
     }
 }
