+++ app/femr/data/models/mysql/concepts/ConceptMedicationForm.java
@@ -16,30 +16,32 @@
      along with fEMR.  If not, see <http://www.gnu.org/licenses/>. If
      you have any questions, contact <info@teamfemr.org>.
 */
-package femr.data.models.mysql;
+package femr.data.models.mysql.concepts;
 
-import femr.data.models.core.IConceptDiagnosis;
+import femr.data.models.core.IConceptMedicationForm;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
 
+//example of medication form: name: caps, description: capsules (predefined in the table)
 @Entity
-@Table(name = "concept_diagnoses")
-public class ConceptDiagnosis implements IConceptDiagnosis {
-
+@Table(name = "concept_medication_forms")
+public class ConceptMedicationForm implements IConceptMedicationForm {
     @Id
     @Column(name = "id", unique = true, nullable = false)
     private int id;
-    @Column(name = "name", nullable = false)
+    @Column(name = "name", unique = true, nullable = false)
     private String name;
+    @Column(name="description")
+    private String description;
+    @Column(name="isDeleted")
+    private Boolean isDeleted;
 
     @Override
     public int getId() {
         return id;
-    }
-
-    @Override
-    public void setId(int id) {
-        this.id = id;
     }
 
     @Override
@@ -50,5 +52,25 @@ public class ConceptDiagnosis implements IConceptDiagnosis {
     @Override
     public void setName(String name) {
         this.name = name;
+    }
+
+    @Override
+    public String getDescription() {
+        return description;
+    }
+
+    @Override
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Override
+    public Boolean getIsDeleted() {
+        return isDeleted;
+    }
+
+    @Override
+    public void setIsDeleted(Boolean isDeleted) {
+        this.isDeleted = isDeleted;
     }
 }
