+++ app/femr/data/models/mysql/concepts/ConceptPrescriptionAdministration.java
@@ -16,7 +16,7 @@
      along with fEMR.  If not, see <http://www.gnu.org/licenses/>. If
      you have any questions, contact <info@teamfemr.org>.
 */
-package femr.data.models.mysql;
+package femr.data.models.mysql.concepts;
 
 import femr.data.models.core.IConceptPrescriptionAdministration;
