+++ app/femr/util/startup/DatabaseSeeder.java
@@ -23,6 +23,10 @@ import femr.data.daos.IRepository;
 import femr.data.daos.Repository;
 import femr.data.models.core.*;
 import femr.data.models.mysql.*;
+import femr.data.models.mysql.concepts.ConceptDiagnosis;
+import femr.data.models.mysql.concepts.ConceptMedicationForm;
+import femr.data.models.mysql.concepts.ConceptMedicationUnit;
+import femr.data.models.mysql.concepts.ConceptPrescriptionAdministration;
 import femr.util.calculations.dateUtils;
 import femr.util.encryptions.BCryptPasswordEncryptor;
 import femr.util.encryptions.IPasswordEncryptor;
